third_party/WebKit/LayoutTests/fast/canvas/canvas-overloads-setShadow-expected.txt

Test the behavior of CanvasRenderingContext2D.setShadow() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.setShadow() threw exception TypeError: Failed to execute 'setShadow' on 'CanvasRenderingContext2D': 3 arguments required, but only 0 present..
PASS ctx.setShadow(0) threw exception TypeError: Failed to execute 'setShadow' on 'CanvasRenderingContext2D': 3 arguments required, but only 1 present..
PASS ctx.setShadow(0, 0) threw exception TypeError: Failed to execute 'setShadow' on 'CanvasRenderingContext2D': 3 arguments required, but only 2 present..
PASS ctx.setShadow(0, 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0, 0, 0) threw exception TypeError: Type error.
PASS ctx.setShadow(0, 0, 0, 0, 'red') is undefined
PASS ctx.setShadow(0, 0, 0, 0, 'red', 0) threw exception TypeError: Type error.
PASS ctx.setShadow(0, 0, 0, 0, 'red', 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0, 0, 0) threw exception TypeError: Type error.
PASS ctx.setShadow(0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0, 0, 0, 0, 0) is undefined
PASS ctx.setShadow(0, 0, 0, 0, 0, 0, 0, 0, 0) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
